Do unsigned shifts.
To quote C99 section 6.5.7 "Bitwise shift operators":

	The result of E1 << E2 is E1 left-shifted E2 bit positions;
	vacated bits are filled with zeros.  ...  If E1 has a signed
	type and nonnegative value, and E1 x 2^E2 is representable in
	the result type, then that is the resulting value; otherwise,
	the behavior is undefined.

In practice, this is harmless, as long as we don't get a trap of some
sort in the "isn't representable" case, but we might as well do an
unsigned shift - it's just doing hashing, so as long as all the relevant
bits get hashed in, that's good enough.

This should keep UBSAN quiet.

+++ optimize.c
@@ -661,7 +661,7 @@ F(opt_state_t *opt_state, int code, int v0, int v1)
 	int val;
 	struct valnode *p;
 
-	hash = (u_int)code ^ (v0 << 4) ^ (v1 << 8);
+	hash = (u_int)code ^ ((u_int)v0 << 4) ^ ((u_int)v1 << 8);
 	hash %= MODULUS;
 
 	for (p = opt_state->hashtbl[hash]; p; p = p->next)
